+++ src/main/java/net/frozenblock/wilderwild/worldgen/modification/WWVegetationGeneration.java
@@ -446,17 +446,6 @@ public final class WWVegetationGeneration {
 				generationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, WWPlacedFeatures.TALL_GRASS.getKey());
 			});
 
-		BiomeModifications.create(WWConstants.id("replace_plains_grass")).add(
-			ModificationPhase.REPLACEMENTS,
-			BiomeSelectors.tag(WWBiomeTags.PLAINS_GRASS),
-			context -> {
-				if (!WWWorldgenConfig.GRASS_GENERATION.get()) return;
-				final BiomeModificationContext.GenerationSettingsContext generationSettings = context.getGenerationSettings();
-				generationSettings.removeFeature(VegetationPlacements.PATCH_GRASS_PLAIN);
-				generationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, WWPlacedFeatures.PATCH_GRASS_PLAINS.getKey());
-				generationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, WWPlacedFeatures.TALL_GRASS_PLAINS.getKey());
-			});
-
 		BiomeModifications.create(WWConstants.id("grass_generation")).add(
 			ModificationPhase.ADDITIONS,
 			BiomeSelectors.all(),